+++ app/pdb/context-cmds.c
@@ -753,7 +753,7 @@ context_get_line_dash_pattern_invoker (GimpProcedure         *procedure,
                                        GError               **error)
 {
   GimpValueArray *return_vals;
-  gint num_dashes = 0;
+  gsize num_dashes = 0;
   gdouble *dashes = NULL;
 
   GimpStrokeOptions *options =
@@ -778,7 +778,7 @@ context_set_line_dash_pattern_invoker (GimpProcedure         *procedure,
                                        GError               **error)
 {
   gboolean success = TRUE;
-  gint num_dashes;
+  gsize num_dashes;
   const gdouble *dashes;
 
   dashes = gimp_value_get_float_array (gimp_value_array_index (args, 0), (gsize *) &num_dashes);
