+++ libgimp/gimpgradient_pdb.h
@@ -38,13 +38,13 @@ gint          gimp_gradient_get_number_of_segments              (GimpGradient   
 gboolean      gimp_gradient_get_uniform_samples                 (GimpGradient              *gradient,
                                                                  gint                       num_samples,
                                                                  gboolean                   reverse,
-                                                                 gint                      *num_color_samples,
+                                                                 gsize                     *num_color_samples,
                                                                  gdouble                  **color_samples);
 gboolean      gimp_gradient_get_custom_samples                  (GimpGradient              *gradient,
                                                                  gint                       num_samples,
                                                                  const gdouble             *positions,
                                                                  gboolean                   reverse,
-                                                                 gint                      *num_color_samples,
+                                                                 gsize                     *num_color_samples,
                                                                  gdouble                  **color_samples);
 GeglColor*    gimp_gradient_segment_get_left_color              (GimpGradient              *gradient,
                                                                  gint                       segment);
